+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Tweaked the discord webhooks layout in settings to include horizontal lines for better clarity.
 * Unfortunately had to remove the functionality to log CMDRs scanned while in a dropship / taxi as we can no longer get the CMDR name from the game journal.
 * The layout of the CMDR information panel in the CMDRs window has been tidied up.
+* Only enable the 'Post to Discord' button on all windows if a Discord username is set in the settings.
 
 ### Bug Fixes: